% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr2sbst.R
\name{chr2sbst}
\alias{chr2sbst}
\title{Convert a chromosome to a subset}
\usage{
chr2sbst(p0)
}
\arguments{
\item{p0}{A string input}
}
\value{
A vector of indices corresponding to \code{p0}
}
\description{
Takes as input a binary string of 0s and 1s, and converts this to a subset for indexing
}
\examples{
x <- '0011001'
chr2sbst(x)
[1] 3 4 7
}
