% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Cross two chromosomes}
\usage{
crossover(p0, p1)
}
\arguments{
\item{p0}{First chromosome}

\item{p1}{Second chromosome}
}
\value{
Returns a child string in which a section of p0 has been replaced by the corresponding section in p1
}
\description{
Simulates a random genetic crossover event between 2 binary chromosomes
}
\examples{
## Example 1
p0 <- '0000000'
p1 <- '1111111'
crossover(p0, p1)
[1] "0001110" #<--This corresponds to a crossover occuring from the 4th through 6th loci

## Example 2
p0 <- '0101'
p1 <- '00001101010'
crossover(p0, p1)
Error: L0 == L1 is not TRUE
}
