% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert a chromosome}
\usage{
invert(p0)
}
\arguments{
\item{p0}{Binary string of 0s and 1s}
}
\value{
Returns the reverse of \code{p0}
}
\description{
Takes a binary chromsome as input and flips the ordering
}
\examples{
x <- '001101'
invert(x)
[1] "101100"
}
