% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate a chromosome}
\usage{
mutate(p0, m.rate)
}
\arguments{
\item{p0}{Binary string of 0s and 1s}

\item{mut.rate}{Mutation rate}
}
\value{
Returns p0 in which some loci have been randomly switched depending on the mutation rate
}
\description{
Creates point mutations in a chromosome given a mutation rate
}
\examples{
x <- '00000000'
m.rate <- 0.01

## Example 1
mutate(x, m.rate)
[1] "00000000" #<--No mutation occurred

## Example 2
mutate(x, m.rate)
[1] "00100000" #<--Mutation occurred at the 3rd locus
}
