% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next.gen.R
\name{next.gen}
\alias{next.gen}
\title{Create the next generation of chromosomes}
\usage{
next.gen(F0, fitness, m.rate)
}
\arguments{
\item{F0}{A list containing a set of binary strings}

\item{fitness}{A vector with the same length as \code{F0} containing the fitness values}
}
\value{
Returns a list with the same dimensions as \code{F0} whose entries correspond to the chromosomes of the next generation of models
}
\description{
Generate a subsequent population of chromosomes given a current population and their fitness values
}
\examples{
F0 <- c('00101', '11100', '01000')
fitness <- c(220, 180, 200)
m.rate <- 0.01
next.gen(F0, fitness, m.rate)
[1] "11100" "10000" "11101" #<--NB: the 'elite' model (i.e., model with the best fitness) reappears in the first position
}
