% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssga.R
\name{plot.ssga}
\alias{plot.ssga}
\title{SSGA violin plot}
\usage{
\method{plot}{ssga}(ssga.obj, B = NULL, nsamples = NULL,
  var.names = NULL, color.flag = FALSE)
}
\arguments{
\item{ssga.obj}{An object of class 'ssga' created using \code{ssga}}

\item{B}{Construct the violin plot using the top \code{B} models}

\item{nsamples}{Total number of samples to use when creating violin plot for each coefficient}

\item{var.names}{Labels to use for the violin plot; by default, column names from the subsettable data are used}

\item{color.flag}{A TRUE/FALSE flag indicating whether to create the violin plot in color or bw}
}
\description{
Create a violin plot based on an input object of class type 'ssga'
}
\examples{
## Load ICU workspace (from aplore3 package)
load(system.file('extdata', 'ssga_icu_ws.RData', package = 'ssga'))
plot(ssga.obj, B = 100, nsamples = 1000, color.flag = TRUE)

## Load Swiss workspace
rm(list = ls(all = TRUE))
load(system.file('extdata', 'ssga_swiss_ws.RData', package = 'ssga'))
plot(ssga.obj)
}
