% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{populate}
\alias{populate}
\title{Populate}
\usage{
populate(pop.size, n.genes)
}
\arguments{
\item{pop.size}{The number of chromosomes to generate}

\item{n.loci}{The number of loci on each chromosome, corresponding to the number of features}
}
\value{
Returns a list of \code{pop.size} randomly generated \code{n.loci}-length chromosomes
}
\description{
Create an initial population of chromosomes
}
\examples{
pop.size <- 2
n.loci <- 10
populate(pop.size, n.loci)
[1] "0101101000" "1100010100"
}
