% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbst2chr.R
\name{sbst2chr}
\alias{sbst2chr}
\title{Convert a subset to a chromosome}
\usage{
sbst2chr(sbst, k)
}
\arguments{
\item{sbst}{A vector of indices}

\item{k}{The desired length of the chromosome}
}
\value{
Returns a binary string of length \code{k} corresponding to which features are present or absent in the model
}
\description{
Takes a vector of indices and converts it to a chromosome
}
\examples{
x <- c(1, 2, 5, 7)
k <- 9
sbst2chr(x, k)
[1] "110010100"
}
