% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssga.freq.R
\name{ssga.freq}
\alias{ssga.freq}
\title{SSGA frequency analysis}
\usage{
ssga.freq(ssga.obj, B = NULL, var.names = NULL)
}
\arguments{
\item{ssga.obj}{An object of class 'ssga' created using \code{ssga}}

\item{B}{Obtain frequencies using the top \code{B} models}
}
\description{
Relative frequencies of variables in an object of class type 'ssga'
}
\examples{
## Load ICU workspace (from aplore3 package)
load(system.file('extdata', 'ssga_icu_ws.RData', package = 'ssga'))
ssga.freq(ssga.obj, B = 500)

## Load Swiss workspace
rm(list = ls(all = TRUE))
load(system.file('extdata', 'ssga_swiss_ws.RData', package = 'ssga'))
ssga.freq(ssga.obj, B = 10)
}
