% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssga.reg.R
\name{ssga.reg}
\alias{ssga.reg}
\title{Shotgun Stochastic Genetic Algorithm}
\usage{
ssga.reg(X, y, no.trials, model.type = "linear", fitness.fun = "AIC",
  forced.vars = NULL, print.flag = TRUE, penalty = NULL)
}
\arguments{
\item{X}{A dataset whose rows represent a vector of p measurements across a set of independent variables}

\item{y}{A set of outcomes}

\item{model.type}{Specify the type of regression model ('linear', 'logistic', 'poisson')}

\item{fitness.fun}{Specify the fitness function ('AIC', 'BIC', 'CUSTOM')}

\item{forced.vars}{A set of 1 or more columns with the same number of rows as \code{X}, representing variables that are to be included in every regression model}

\item{print.flag}{When \code{print.flag} is set to TRUE, the current trial, iteration, best model fitness score, and chromosome corresponding to the best model will be shown in the console}

\item{penalty}{If \code{fitness.fun = 'CUSTOM'}, the fitness function takes the form -2*logLike + \code{penalty}*nparam, where logLike refers to the maximized log-likelihood of the regression model and nparam is the total number of free parameters in the model; note: AIC = -2*logLike + 2*nparam}
}
\value{
Returns an object of class 'ssga'
}
\description{
Simultaneously perform variable subet selection and model averaging with the shotgun stochastic genetic algorithm regression
}
\examples{
## Load ICU workspace for logistic regression (data were obtained from aplore3 package)
load(system.file('extdata', 'ssga_icu_ws.RData', package = 'ssga'))
no.trials <- 5
model.type <- 'logistic'
fitness.fun <- 'AIC'
ssga.obj <- ssga.reg(X, y,
                     no.trials = no.trials,
                     model.type = model.type,
                     fitness.fun = fitness.fun,
                     print.flag = TRUE)
summary(ssga.obj)


## Poisson regression with the ICU data
model.type <- 'poisson'
y <- X[ , 'age']
X <- X[ , !colnames(X)\%in\%'age']
ssga.obj <- ssga.reg(X, y,
                     no.trials = no.trials,
                     model.type = model.type,
                     fitness.fun = fitness.fun)
summary(ssga.obj)


## Load Swiss workspace for linear regression
rm(list = ls(all = TRUE))
load(system.file('extdata', 'ssga_swiss_ws.RData', package = 'ssga'))
no.trials <- 5
model.type <- 'linear'
fitness.fun <- 'AIC'
ssga.obj <- ssga.reg(X, y,
                     no.trials = no.trials,
                     model.type = model.type,
                     fitness.fun = fitness.fun,
                     print.flag = TRUE)
summary(ssga.obj)
}
