% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.nbd.R
\name{sss.nbd}
\alias{sss.nbd}
\title{Shotgun Stochastic Search Neighborhood Generator}
\usage{
sss.nbd(p0, token)
}
\arguments{
\item{p0}{Binary string of 0s and 1s}

\item{token}{A character which is either '+', '-', or 'o', depending on which neighborhood is to be generated}
}
\value{
Returns a list containing all chromosomes making up the given neighborhood requested using \code{token}.  Let k denote the number of 1s in \code{p0}, and k' denote the number of 0s.  If \code{token == '+'}, \code{nbd} returns the set of length k' in which each chromosome is sequentially \code{p0} augmented by one of its 0s switched to a 1; if \code{token == '-'}, \code{nbd} returns the set of length k in which each chromsome is sequentially \code{p0} decreased by one of its 1s switched to a 0; if \code{token == 'o'}, returns the set of length k*k' in which each chromosome is simultaneously \code{p0} decreased by one of its 1s switched to a 0 while each of its 0s is switched to a 1.
}
\description{
Create the 3 neighborhoods from a shotgun stochastic search algorithm (see Hans, et al. (2007))
}
\examples{
p0 <- '00101'
nbd(p0, '+')
[1] "10101" "01101" "00111"
nbd(p0, '-')
[1] "00001" "00100"
nbd(p0, 'o')
[1] "10001" "01001" "00011" "10100" "01100" "00110"
nbd(p0, 'z')
Error: (token == "+") || (token == "-") || (token == "o") is not TRUE 
}
